requirements/base.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
requests>=2.25.1
celery>=5.0.5
redis>=3.5.3
psycopg2-binary>=2.8.6
